+++ app/Http/Controllers/PostController.php
@@ -2,9 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\PostFileAddRequest;
+use App\Http\Requests\PostFileDelRequest;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
+// use App\Http\Requests\CommentStoreRequest;
+// use App\Http\Requests\LikeRequest;
+use App\Http\Requests\ListRequest;
 use App\Models\Post;
+use App\Transformers\PostTransformer;
+// use App\Transformers\CommentTransformer;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Validator;
+use Spatie\Fractal\Facades\Fractal;
 
 class PostController extends Controller
 {
@@ -14,57 +26,215 @@ class PostController extends Controller
     }
 
     /**
-     * Display a listing of the resource.
+     * 列表
      *
-     * @return \Illuminate\Http\Response
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function index()
+    public function index(Request $request): JsonResponse
     {
-        //
+        Validator::make($request->all(), [
+            'page' => 'sometimes|numeric|min:1',
+            'limit' => 'sometimes|numeric|min:1|max:100',
+            'sort' => 'sometimes|in:id_asc,id_desc,updated_at_asc,updated_at_desc',
+        ])->validate();
+
+        $page = $request->input('page', 1);
+        $limit = $request->input('limit', 5);
+        $sort = $request->input('sort', 'id_asc');
+
+        [$column, $order] = preg_split('/_(?=(asc|desc)$)/', $sort);
+
+        $posts = $request->user()
+            ->posts()
+            ->with(['tags', 'loveReactant.reactions', 'loveReactant.reactionCounters'])
+            ->orderBy($column, $order)
+            ->paginate($limit, ['*'], 'page', $page);
+
+        return Fractal::create($posts, new PostTransformer())
+            // * 手動includes
+            // ->parseIncludes('tag')
+            // ->includeTags()
+            ->respond();
     }
 
     /**
-     * Store a newly created resource in storage.
+     * 新增
      *
      * @param  \App\Http\Requests\StorePostRequest  $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function store(StorePostRequest $request)
+    public function store(StorePostRequest $request): JsonResponse
     {
-        //
+        $post = $request->user()->posts()->create($request->all());
+
+        if ($request->has('tag')) {
+            $post->setTag($request->input('tag') ?? []);
+        }
+
+        $post->setFile('gallery', $request->input('gallery') ?? []);
+
+        return Fractal::create($post, new PostTransformer())
+            // * 手動includes
+            // ->parseIncludes('tag')
+            // ->includeTags()
+            ->respond();
     }
 
     /**
-     * Display the specified resource.
+     * 顯示
      *
      * @param  \App\Models\Post  $post
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function show(Post $post)
+    public function show(Post $post): JsonResponse
     {
-        //
+        return Fractal::create($post, new PostTransformer())
+            // * 手動includes
+            // ->parseIncludes('tag')
+            // ->includeTags()
+            ->respond();
     }
 
     /**
-     * Update the specified resource in storage.
+     * 更新
      *
      * @param  \App\Http\Requests\UpdatePostRequest  $request
      * @param  \App\Models\Post  $post
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function update(UpdatePostRequest $request, Post $post)
+    public function update(UpdatePostRequest $request, Post $post): JsonResponse
     {
-        //
+        $post->update($request->all());
+
+        if ($request->has('tag')) {
+            $post->setTag($request->input('tag') ?? []);
+        }
+
+        return Fractal::create($post, new PostTransformer())
+            // * 手動includes
+            // ->parseIncludes('tag')
+            // ->includeTags()
+            ->respond();
     }
 
     /**
-     * Remove the specified resource from storage.
+     * 刪除
      *
      * @param  \App\Models\Post  $post
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Post $post)
+    public function destroy(Post $post): Response
     {
-        //
+        return response($post->delete(), 200);
+    }
+
+    /**
+     * 檔案新增
+     *
+     * @param PostFileAddRequest $request
+     * @param \App\Models\Post  $post
+     * @return JsonResponse
+     */
+    public function fileAdd(PostFileAddRequest $request, Post $post): JsonResponse
+    {
+        $post->setFile($request->input('collection'), [$request->input('file')]);
+
+        return Fractal::create($post, new BlogTransformer())
+            ->parseIncludes($request->input('collection'))
+            ->respond();
+    }
+
+    /**
+     * 檔案刪除
+     *
+     * @param PostFileDelRequest $request
+     * @param \App\Models\Post  $post
+     * @return JsonResponse
+     */
+    public function fileDel(PostFileDelRequest $request, Post $post): JsonResponse
+    {
+        // 配合 PostFileDelRequest 檢查 media_id 是否存在於資料表
+        // $mediaItems = $post->getMedia($request->input('collection'));
+
+        // $mediaItem = $mediaItems->find($request->input('media_id'));
+
+        // if($mediaItem){
+        //     return response()->json(['message' => 'done']);
+        // }else{
+        //     return response()->json(['message' => 'media not found.'], 404);
+        // }
+
+        // 補捉model錯誤訊息
+        try {
+            $post->delFile($request->input('collection'), $request->input('media_id'));
+        } catch (\Exception $e) {
+            return response()->json(['message' => $e->getMessage()], 404);
+        }
+
+        return response()->json(['message' => 'done']);
+    }
+
+    /**
+     * 評論列表
+     *
+     * @param \App\Http\Requests\ListRequest $request
+     * @param \App\Models\Post  $post
+     * @return JsonResponse
+     */
+    public function comments(ListRequest $request, Post $post): JsonResponse
+    {
+        $page = $request->input('page', 1);
+        $limit = $request->input('limit', 5);
+        $sort = $request->input('sort', 'id_asc');
+
+        [$column, $order] = preg_split('/_(?=(asc|desc)$)/', $sort);
+
+        $comments = $post->comments()
+            ->approved()
+            ->orderBy($column, $order)
+            ->paginate($limit, ['*'], 'page', $page);
+
+        return Fractal::create($comments, new CommentTransformer())
+            ->respond();
+    }
+
+    /**
+     * 評論新增
+     *
+     * @param CommentStoreRequest $request
+     * @param \App\Models\Post  $post
+     * @return JsonResponse
+     */
+    public function commentCreate(CommentStoreRequest $request, Post $post): JsonResponse
+    {
+        $comment = $post->comment($request->input('comment'));
+
+        return Fractal::create($comment, new CommentTransformer())
+            ->respond();
+    }
+
+    /**
+     * 設定喜歡或不喜歡
+     *
+     * @param LikeRequest $request
+     * @param int $id
+     * @return void
+     */
+    public function like(LikeRequest $request, int $id): JsonResponse
+    {
+        // ? like_count,dislike_count數字不同步問題
+        // TODO:修改.env的QUEUE_CONNECTION=sync才會同步
+        // TODO:思考該不該顯示數字
+
+        $post = Post::with([
+            'loveReactant.reactions',
+        ])->find($id);
+
+        $request->user()
+            ->setLike($post, $request->input('type', ''));
+
+        return Fractal::create($post->fresh(), new PostTransformer())
+            ->respond();
     }
 }