+++ app/Transformers/TagTransformer.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace App\Transformers;
+
+use League\Fractal\TransformerAbstract;
+use Spatie\Tags\Tag;
+
+class TagTransformer extends TransformerAbstract
+{
+    /**
+     * List of resources to automatically include
+     *
+     * @var array
+     */
+    protected array $defaultIncludes = [
+        //
+    ];
+
+    /**
+     * List of resources possible to include
+     *
+     * @var array
+     */
+    protected array $availableIncludes = [
+        //
+    ];
+
+    /**
+     * A Fractal transformer.
+     *
+     * @return array
+     */
+    public function transform(Tag $tag)
+    {
+        return [
+            'id' => (int) $tag->id,
+            'name' => $tag->name,
+            'type' => $tag->type,
+        ];
+    }
+}